requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
flask-session==0.5.0